@@ -1,4 +1,4 @@
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import appContext from "../stores/appContext";
 import Provider from "../labs/Provider";
 import appStore from "../stores/appStore";
@@ -39,6 +39,7 @@ export function showDialog<T extends DialogProps>(
   props?: Omit<T, "destroy">
 ): DialogCallback {
   const tempDiv = document.createElement("div");
+  const dialog = createRoot(tempDiv);
   document.body.append(tempDiv);
 
   setTimeout(() => {
@@ -50,8 +51,8 @@ export function showDialog<T extends DialogProps>(
       tempDiv.firstElementChild?.classList.remove("showup");
       tempDiv.firstElementChild?.classList.add("showoff");
       setTimeout(() => {
+        dialog.unmount();
         tempDiv.remove();
-        ReactDOM.unmountComponentAtNode(tempDiv);
       }, ANIMATION_DURATION);
     },
   };
@@ -75,7 +76,7 @@ export function showDialog<T extends DialogProps>(
     );
   }
 
-  ReactDOM.render(Fragment, tempDiv);
+  dialog.render(Fragment);
 
   return cbs;
 }
